Squashed 'fast-downward-conjunctions/' changes from a4db4c8..a7f3ab1
a7f3ab1 added march=native and PGO builds
f5b4abe removed misc, experiments, and other unused files
cafb951 removed TODOs and unused conflict extraction algorithm
f9b3017 fixed compile error
61a4331 Revert "estimate fewest counters if too expensive"
5dfea5e ipc18 cleanup
ca36e7a die gracefully if the task has conditional effects
fdff320 don't ignore the cost bound in grey initial state search
5dc76a0 fixed successor generator path
9a558a1 stop grey heuristic evaluation early if the plan cannot be repaired
08b371c added grey painting option
0b166f7 skip expensive memory allocation in connectivity check if there are no ops from one value to another
ff1d1f3 tie break by operator applicability in dtg search
1dd70c6 fixed gcc issues
d6f5232 remove red sufficient unachieved iterators
642e428 don't compare incompatible iterators
f00216f grey planning merge
b0fa454 write to output, not output.sas
c053d9b h2 preprocessor (w/o SG/DTGs)
353afad fixed heuristic deletion bug in ipc18 iterated search
56b1aec allow an arbitrary decision point for skip on fail/solve
fd8fe3c fixed missing heuristic cache argument, store operator indices instead of pointers
7d0d9a1 made search timer accessible for subclasses of search engine, added timeout checks in main ehc and refinement loops
11e2804 fixed missing initialization for heuristic cache option
69bf6f6 minor refactoring
ec4a0d0 novelty heuristics now unsubscribe themselves when deleted
88c5118 improved novelty conjunction linking including a bug fix when the novelty heuristic was deleted before the conjunctions heuristic, minor refactoring
ca68fb5 fixed zero-cost actions in EHC
c8abf9d skip depending on the outcome of the first search
233c59c improved skip and deletion handling
d723975 added iterated weights lazy astar, and ipc18 iterated search as a wrapper
7aeeb61 take the action cost into account when pruning counters
7d77ee8 fixed a bug introduced in revision 1561dcaba68d03ace3af9075ea24e9e0f3a8e65e
42680ee added conflict-based tie breaking, some cleanup
13139f4 removed old tie breaking implementation
41fe63e lazy search cost bound fixes
3085a62 don't switch to counter estimates by default
1561dca don't iterate over all actions all the time
65b2cea fixed initial state solutions... again
afcaa93 updated successor generator
2a0cd90 Move successor_generator to task_utils and give it its own plugin and namespace
051c22e don't build/dump causal graph, domain transition graphs, successor generator
REVERT: a4db4c8 removed misc, experiments, and other unused files
REVERT: aa424e8 removed old tie breaking implementation
REVERT: 9392880 removed TODOs and unused conflict extraction algorithm
REVERT: 5abe1cf Revert "estimate fewest counters if too expensive"
REVERT: 4f7a429 ipc18 cleanup
REVERT: 06b2f82 don't ignore the cost bound in grey initial state search
REVERT: b461556 fixed successor generator path
REVERT: ceb08ce stop grey heuristic evaluation early if the plan cannot be repaired
REVERT: d5e2d9b added grey painting option
REVERT: 081ef52 skip expensive memory allocation in connectivity check if there are no ops from one value to another
REVERT: 639aafe tie break by operator applicability in dtg search
REVERT: edaa640 fixed gcc issues
REVERT: 73ef9dd remove red sufficient unachieved iterators
REVERT: e527c14 don't compare incompatible iterators
REVERT: bd1b876 grey planning merge
REVERT: e46ab78 Merge branch 'master' into h2_preprocessor
REVERT: 7bc75dc lazy search cost bound fixes
REVERT: f5b5d2f Merge branch 'master' into h2_preprocessor
REVERT: 5b0315d don't switch to counter estimates by default
REVERT: a08baee don't iterate over all actions all the time
REVERT: 0eb596c Merge branch 'master' into h2_preprocessor
REVERT: 6b23be0 fixed initial state solutions... again
REVERT: a653549 Merge branch 'master' into h2_preprocessor
REVERT: 29b59d2 write to output, not output.sas
REVERT: ffaf24e h2 preprocessor
REVERT: db06d64 Merge branch 'successor_generator_update'
REVERT: 4c41a53 updated successor generator
REVERT: f26d99c Move successor_generator to task_utils and give it its own plugin and namespace
REVERT: 857d309 don't build and dump causal graph, domain transition graphs, successor generator

git-subtree-dir: fast-downward-conjunctions
git-subtree-split: a7f3ab1cb595d82710c76f488ccaacd206096465

+++ src/search/conjunctions/conflict_extraction.cc
@@ -1063,7 +1063,7 @@ auto ConflictExtraction::ConflictExtractionHelper<ConflictType>::generate_candid
 						}
 						const auto path = construct_path(deleter, failed);
 						assert(!path.empty());
-						assert(path.size() - 1 == priority);
+						assert(static_cast<int>(path.size()) - 1 == priority);
 						// add the pair {p, q_n} as conflict where p is the deleted fact and q is the last label on the path from the deleter to the failed action
 						add_combined_conflict(conflicts[priority], *path.back(), *precondition.first, bsg.nodes[deleter], *first_deleted, deleter - failed, heuristic, online_scoring);
 					} else {
